requirements.txt
python-dotenv
pyyaml
pydantic
numpy
scikit-learn
faiss-cpu
PyPDF2
rich
